Move HtmlOptions interface to schema.ts
